+++ scripts/shared/common-module.ps1
@@ -481,3 +481,18 @@ function Register-ToolboxAutoUpdate {
     $trigger = New-ScheduledTaskTrigger -AtLogon
     Register-ScheduledTask -Action $action -Trigger $trigger -TaskPath "\Toolbox\AutoUpdate" -TaskName $taskName -Description "Start Toolbox auto update" -ErrorAction SilentlyContinue | Out-Null
 }
+
+$rootPath = Resolve-Path -Path "$PSScriptRoot\..\.." -ErrorAction Stop
+[System.Environment]::SetEnvironmentVariable("TOOLBOX_HOME", $($rootPath.Path), "Process")
+[System.Environment]::SetEnvironmentVariable("TOOLBOX_APPS", "$($rootPath.Path)\local\apps", "Process")
+[System.Environment]::SetEnvironmentVariable("TOOLBOX_PLANS", "$($rootPath.Path)\local\plans", "Process")
+[System.Environment]::SetEnvironmentVariable("TOOLBOX_BIN", "$($rootPath.Path)\local\bin", "Process")
+
+$variables = Get-CompanyEnvironmentVariables
+$restrictedVariables = @("HOME", "APPS", "PLANS", "BIN")
+foreach ($variableName in $variables.PSObject.Properties.Name) {
+	if ($variableName -notin $restrictedVariables) {
+		$variableValue = $variables.$variableName
+		[System.Environment]::SetEnvironmentVariable("TOOLBOX_$variableName", $variableValue, "Process")
+	}
+}