+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.& (November 8, 2023)
+
+* Changed the load of the process environment variables from toolbox.ps1 to the shared common module
+* The shared common now load the environment variables so plans CLI can load those variables as well
+
 # 1.0.0 (November 7, 2023)
 
 * First official release of Toolbox